+++ app/Http/Controllers/SecondProductController.php
@@ -45,7 +45,7 @@ class SecondProductController extends Controller
      */
     public function create()
     {
-        //
+        abort(404);
     }
 
     /**
@@ -99,7 +99,7 @@ class SecondProductController extends Controller
      */
     public function show($id)
     {
-        //
+        abort(404);
     }
 
     /**
@@ -110,7 +110,7 @@ class SecondProductController extends Controller
      */
     public function edit($id)
     {
-        //
+        abort(404);
     }
 
     /**
@@ -122,6 +122,9 @@ class SecondProductController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if (!Gate::allows('level1') && !Gate::allows('level2')) {
+            abort(403);
+        }
         $validate_data = $request->validate([
             "stock_" => "required|numeric",
 
@@ -141,6 +144,9 @@ class SecondProductController extends Controller
      */
     public function destroy($id)
     {
+        if (!Gate::allows('level1') && !Gate::allows('level2')) {
+            abort(403);
+        }
         $model = SecondProductModel::find($id);
         $model->delete();
         return redirect('/second_product')->with('error', 'Delete Data Stocks Success');
